% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson-siegel.R
\name{beta1_regressor}
\alias{beta1_regressor}
\title{Beta_1 Regressor}
\usage{
beta1_regressor(maturity, lambda = 0.4)
}
\arguments{
\item{maturity}{A number. Should be represented as years, e.g. 3 months = 0.25 years.}

\item{lambda}{A number between 0 and 1. Defaults to 0.4 as is given by the Academy's generator.}
}
\value{
A numeric vector giving the Beta_1 regressor for the Nelson-Siegel Yield Curve interpolation.
}
\description{
Beta_1 Regressor
}
\examples{
beta1_regressor(1)
beta1_regressor(1,0.5)
beta1_regressor(c(1,20))
}
