% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson-siegel.R
\name{fit_parameters}
\alias{fit_parameters}
\title{Get Nelson-Siegel Factors}
\usage{
fit_parameters(b, A)
}
\arguments{
\item{short_rate}{The one year interest rate.}

\item{long_rate}{The twenty year interest rate.}
}
\value{
b_0 and b_1 factors are calculated by solving A*x=b for x.
}
\description{
The AIRG assumes a short rate maturity of one year and a long rate maturity of twenty years.
}
\examples{
get_factors(0.0263,0.0287)
}
