% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_processes}
\alias{generate_processes}
\title{Generate processes.}
\usage{
generate_processes(N, short_rate = 0.0263, long_rate = 0.0287,
  beta_i = 0.00509, beta_alpha = 0.02685, beta_nu = 0.04001,
  tau_i = 0.035, tau_alpha = 0.01, tau_nu = 0.0287,
  sigma_i = 0.0287, sigma_alpha = 0.04148, sigma_nu = 0.11489,
  phi = 2e-04, psi = 0.25164, theta = 1)
}
\value{
A matrix with rows; 1 = log of long rate, 2 = difference (long - short), 3 = log of long rate volatility
}
\description{
Generate processes.
}
