% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_rates}
\alias{generate_rates}
\title{Generate rates based on processes}
\usage{
generate_rates(N)
}
\value{
A 2xn matrix with rows: 1 = short rate, 2 = long rate.
}
\description{
Generate rates based on processes
}
