% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastics.R
\name{generate_stochastics}
\alias{generate_stochastics}
\title{Generate Stochastics}
\usage{
generate_stochastics(months, processes = 3)
}
\arguments{
\item{months}{An integer of the number of projection months.}

\item{processes}{Defaults to three stochastic processes.}
}
\value{
An MxN matrix of stochastic random variables where M = months and N = processes
}
\description{
Generate Stochastics
}
\examples{
generate_stochastics(360,3)
}
