% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson-siegel.R
\name{get_rates}
\alias{get_rates}
\title{Get Rates}
\usage{
get_rates(factors, regressors)
}
\arguments{
\item{factors}{A matrix.}

\item{regressors}{A matrix.}
}
\value{
A matrix of interest rates
}
\description{
Get Rates
}
\examples{
get_rates(matrix(c(1,1,0.8241999,0.1249581),nrow=2),matrix(c(0.029128892,-0.003432289),nrow=2))
}
