% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson-siegel.R
\name{ns_regressors}
\alias{ns_regressors}
\title{Nelson-Siegel Regressors}
\usage{
ns_regressors(maturities)
}
\arguments{
\item{maturities}{A number or numeric vector. Should be represented as years, e.g. 3 months = 0.25 years.}
}
\value{
A matrix with length as number of maturities of Nelson-Siegel regressors of the form: \begin{bmatrix}1 \Beta_1 \end{bmatrix}
}
\description{
Nelson-Siegel Regressors
}
\examples{
ns_regressors(c(0.25,0.5,1,5,10,20,30))
}
