% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastics.R
\name{correlate_stochastics}
\alias{correlate_stochastics}
\title{Correlate Stochastics}
\usage{
correlate_stochastics(stochastics, correlation)
}
\arguments{
\item{stochastics}{}

\item{correlation}{}
}
\value{
A correlated MxN matrix of stochastic random variables
}
\description{
Correlate Stochastics
}
